requirements/base.txt
google-cloud-aiplatform==1.68.0
google-cloud-pipeline-components==2.8.0
kfp==2.7.0
typer==0.12.5